[#40] Feat : PriceSlider 컴포넌트를 PriceModal에 병합 & 초기화 기능 연동
